refactor: split test files to match module boundaries
- animation.test.ts now only tests createAnimationLoop (from animation.ts)
- dom.test.ts tests createGifPopup, createFallingItems, spawnAnimatedSprite (from dom.ts)
